+++ src/di/tests/ContainerTest.php
@@ -1,7 +1,16 @@
 <?php
 
-namespace HyperfTest\Di;
+declare(strict_types=1);
+/**
+ * This file is part of Hyperf.
+ *
+ * @link     https://www.hyperf.io
+ * @document https://doc.hyperf.io
+ * @contact  group@hyperf.io
+ * @license  https://github.com/hyperf-cloud/hyperf/blob/master/LICENSE
+ */
 
+namespace HyperfTest\Di;
 
 use Hyperf\Di\Annotation\Scanner;
 use Hyperf\Di\Container;
@@ -10,9 +19,12 @@ use HyperfTest\Di\Stub\Foo;
 use HyperfTest\Di\Stub\FooInterface;
 use PHPUnit\Framework\TestCase;
 
+/**
+ * @internal
+ * @coversNothing
+ */
 class ContainerTest extends TestCase
 {
-
     public function testHas()
     {
         $container = new Container(new DefinitionSource([], [], new Scanner()));
@@ -20,5 +32,4 @@ class ContainerTest extends TestCase
         $this->assertFalse($container->has(NotExistClass::class));
         $this->assertTrue($container->has(Foo::class));
     }
-
 }